install postcss with tailwindcss, add example directory
