+++ example/main.jsx
@@ -1,0 +1,26 @@
+import React, {useState} from 'react'
+import ReactDOM from 'react-dom'
+import '../src/assets/tailwind.css';
+import { TextField } from '../src/index';
+
+function App() {
+  const [textValue, setTextValue] = useState('');
+  return (
+    <div className="p-10 bg-red-300">
+      Components:
+      <ul>
+        <li>
+          <p>Textfield component</p>
+          <TextField value={textValue} onChange={(event) => setTextValue(event.target.value)} />
+        </li>
+      </ul>
+    </div>
+  );
+};
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+)
