+++ package.json
@@ -16,13 +16,14 @@
     }
   },
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview"
+    "dev": "vite example",
+    "build": "vite build"
   },
   "dependencies": {
+    "autoprefixer": "^10.2.4",
     "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react-dom": "^17.0.0",
+    "tailwindcss": "^2.0.3"
   },
   "devDependencies": {
     "@vitejs/plugin-react-refresh": "^1.3.1",
