+++ src/assets/tailwind.css
@@ -1,0 +1,7 @@
+@tailwind base;
+
+@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
+
+@tailwind components;
+
+@tailwind utilities;
