+++ src/index.js
@@ -1,0 +1,5 @@
+import TextField from './components/TextField';
+
+export {
+  TextField
+}
